requirements.txt
opencv-python
mysql-connector-python
numpy
ultralytics
shapely
openai
python-dotenv